requirements.txt
flask
pillow
numpy
gtts